+++ src/NSchema.Core/Project/Nsql/Tokens/NsqlKeywords.cs
@@ -19,6 +19,7 @@ internal static class NsqlKeywords
     public const string Begin = "BEGIN";
     public const string Cache = "CACHE";
     public const string Cascade = "CASCADE";
+    public const string Restrict = "RESTRICT";
     public const string Check = "CHECK";
     public const string Column = "COLUMN";
     public const string Clustered = "CLUSTERED";
@@ -92,6 +93,9 @@ internal static class NsqlKeywords
     public const string Start = "START";
     public const string Statement = "STATEMENT";
     public const string Stored = "STORED";
+    public const string Virtual = "VIRTUAL";
+    public const string RowGuidCol = "ROWGUIDCOL";
+    public const string Replication = "REPLICATION";
     public const string Table = "TABLE";
     public const string Template = "TEMPLATE";
     public const string To = "TO";